works 10% of the time

+++ include/kernel/typedef.h
@@ -2,7 +2,7 @@
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned long int uint64_t;
 typedef unsigned long int size_t;
 typedef uint64_t uintptr_t;
